requirements.txt
sympy==1.9
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
torch==2.0.1
# torch==1.9.1
tqdm==4.66.6
transformers==4.46.1
scikit-learn
ipython
easydict
python-dotenv
sentencepiece
bitsandbytes